Convert shell.cpp to .c; starting to write socket programming.
