+++ category-software/Identify_Yourself/src/Makefile
@@ -1,7 +1,7 @@
 EXEC_NAME=run_me
 
-all: shell.cpp
-	gcc -lstdc++ -o $(EXEC_NAME) shell.cpp
+all: shell.c
+	gcc -o $(EXEC_NAME) shell.c
 
 clean:
 	rm -f $(EXEC_NAME)
